requirements.txt
redis==4.5.4
celery==5.2.7
requests==2.28.2
python-dotenv==1.0.0

